Will move day00 files to a folder
